+++ Variant/Setup/UpgradeSchema.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Pimgento\Variant\Setup;
+
+use Magento\Framework\Setup\UpgradeSchemaInterface;
+use Magento\Framework\Setup\ModuleContextInterface;
+use Magento\Framework\Setup\SchemaSetupInterface;
+use \Magento\Framework\DB\Ddl\Table;
+
+/**
+ * @codeCoverageIgnore
+ */
+class UpgradeSchema implements UpgradeSchemaInterface
+{
+
+    /**
+     * {@inheritdoc}
+     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
+     */
+    public function upgrade(SchemaSetupInterface $setup, ModuleContextInterface $context)
+    {
+        $installer = $setup;
+
+        $installer->startSetup();
+
+        if (version_compare($context->getVersion(), '1.0.1', '<=')) {
+            $installer->run("ALTER TABLE `" . $installer->getTable('pimgento_variant') . "` ENGINE=MyISAM");
+        }
+
+        $installer->endSetup();
+    }
+}