+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [100.2.21] - 2017-09-25
+
+  * fix: Variant table engine for big data
+  * fix: Attribute attached to a lot of families
+
 ## [100.2.20] - 2017-09-05
 
   * update: disabled media importation if needed